# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  description: string;
  material: string;
  imageUrl: string;
  detail: string;
  season: string;
  available: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  description: string;
  material: string;
  /** @minLength 1 */
  imageUrl: string;
  detail: string;
  season: string;
  available?: boolean;
  sortOrder?: number;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  category?: string;
  description?: string;
  material?: string;
  /** @minLength 1 */
  imageUrl?: string;
  detail?: string;
  season?: string;
  available?: boolean;
  sortOrder?: number;
}

export interface Review {
  id: number;
  displayName: string;
  body: string;
  rating: number;
  approved: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ReviewInput {
  /** @minLength 1 */
  displayName: string;
  /** @minLength 1 */
  body: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
}

export interface ReviewModeration {
  approved: boolean;
}

export interface UploadRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
}