# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  category: string;
  description: string;
  material: string;
  imageUrl: string;
  detail: string;
  season: string;
  available: boolean;
  sortOrder: number;
  createdAt: Date;
  updatedAt: Date;
}