# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  displayName: string;
  body: string;
  rating: number;
  approved: boolean;
  createdAt: Date;
  updatedAt: Date;
}