# lib/api-zod/src/generated/types/reviewInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewInput {
  /** @minLength 1 */
  displayName: string;
  /** @minLength 1 */
  body: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
}